// src/types/disappearance.ts
export type TDisappearance = {
    _id: string,
    img_tagline_bangla: string;
    img_tagline_english: string;
    admin_name: string;
    date: string;
    bangla_title: string;
    english_title: string;
    name_bangla: string;
    name_english: string;
    designation_bangla: string;
    designation_english: string;
    address_bangla: string;
    address_english: string;
    category: string;
    bangla_short_description: string;
    english_short_description: string;
    name_published_newspaper: string;
    news_release_date: string;
    link_published_newspaper: string;
    meta_title: string;
    meta_keywords: string[];
    meta_description: string;
    thumnail_images: string[],
    eng_iamges: string[],
    bng_Images: string[],
};